refactor: move the component to a separate module
