+++ src/App.jsx
@@ -1,14 +1,7 @@
 import { Layout } from 'antd';
 import AppSider from './components/layout/AppSider.jsx';
 import AppHeader from './components/layout/AppHeader.jsx';
-
-const contentStyle = {
-  textAlign: 'center',
-  minHeight: 'calc(100vh - 60px)',
-  lineHeight: '120px',
-  color: '#fff',
-  backgroundColor: '#001529',
-};
+import AppContent from './components/layout/AppContent';
 
 export default function App() {
   return (
@@ -16,7 +9,7 @@ export default function App() {
       <AppHeader />
       <Layout>
         <AppSider />
-        <Layout.Content style={contentStyle}>Content</Layout.Content>
+        <AppContent />
       </Layout>
     </Layout>
   );
